refactor `manual_filter`
Move common functions to `manual_utils.rs`, better arm matching, use clippy utils `contains_unsafe_block`
